requirements.txt
netmiko==4.5.0
python-dotenv==1.0.1
slack_sdk==3.34.0
Flask==3.1.0